requirements.txt
# NextReel Requirements — Production
# ===================================
# Last updated: 2026-03-22
#
# Core Web Framework & Server
# ---------------------------
quart==0.20.0                    # Async web framework (Flask-like API)
hypercorn==0.17.3                # ASGI server for running Quart apps
quart-session==3.0.0             # Session management for Quart
werkzeug==3.1.6                  # WSGI utilities (used by Quart)

# Database & Storage
# -----------------
aiomysql==0.3.2                  # Async MySQL/MariaDB client (pulls PyMySQL transitively)
redis>=4.2.0,<6                  # redis-py; <6 required by arq 0.26.x (server can be Redis 6/7)

# API Clients & HTTP
# -----------------
httpx==0.28.1                    # Modern async HTTP client

# Security & Authentication
# ------------------------
cryptography==46.0.5             # Cryptographic primitives for security features

# Configuration & Environment
# --------------------------
python-dotenv==1.0.1             # Load environment variables from .env file

# Monitoring & Observability
# -------------------------
prometheus-client==0.21.1        # Prometheus metrics collection

# Logging & Debugging (Optional but Recommended)
# ----------------------------------------------
colorlog==6.9.0                  # Colored terminal output for logs

# Task Scheduling & Background Jobs
# ---------------------------------
arq==0.26.3                      # Redis-backed worker for enrichment and refresh jobs

# Utilities
# --------
jinja2==3.1.6                    # Template engine (used by Quart)
markupsafe==3.0.2                # Safe string markup for Jinja2
